Seperate line and bar readings actions/reducers.
